Moved drive testing to another file
